Stop truncating the stream buffer pointer in hwFlushStream
Streamed audio uploads computed their source as (u32)base + offset,
dropping the top half of a host pointer. The result still looked like a
plausible address, so the copy went ahead and read from wherever that
landed -- a fault a long way from the code that caused it, and only once
streaming actually started, which is why it surfaced in a menu rather
than at startup.

The destination bounds check added earlier could not catch this: it
validates the ARAM side of the transfer, and the bad address was the main
memory side.

Also report a refused upload through MUSY_REPORT rather than MUSY_DEBUG,
which compiles to nothing outside debug builds -- a dropped transfer was
silently losing sample data with no way to notice.

Verified over two minutes of driven menu navigation: no faults, and the
mixer produces audio on about a fifth of frames, peaking at 18819.

+++ src/musyx/runtime/hardware.c
@@ -6,6 +6,8 @@
 
 #include "musyx/musyx.h"
 #include "musyx/hardware.h"
+
+#include <stdint.h>
 #include "math.h"
 #include "float.h"
 #include "musyx/assert.h"
@@ -494,7 +496,9 @@ void hwFlushStream(void* base, u32 offset, u32 bytes, u8 hwStreamHandle, void (*
   bytes += (offset & 31);
   offset &= ~31;
   bytes = (bytes + 31) & ~31;
-  mram = (u32)base + offset;
+  /* base is a host pointer: truncating it to 32 bits leaves a source address
+   * that no longer refers to the stream buffer. */
+  mram = (uintptr_t)base + offset;
 #if MUSY_TARGET == MUSY_TARGET_DOLPHIN
   DCStoreRange((void*)mram, bytes);
 #endif
